implements setting of pref values

+++ app/src/main/java/com/tailorfit/android/tailorfitapp/signup/SignUpFragment.kt
@@ -59,7 +59,7 @@ class SignUpFragment : BaseViewModelFragment() {
                         binding.storeEditText.stringContent(),
                         binding.nameEditText.stringContent(),
                         binding.passwordEditText.stringContent(),
-                        binding.phoneEditText.stringContent().toLong() // TODO. Why Long??. Collect phone number as String 🤨
+                        binding.phoneEditText.stringContent() 
                     )
                 )
             }
